MueLu hierarchical: refactor
